WIP: Renaming file 'ContactPictureLoader' with Kotlin extension
