+++ app/ui/src/main/java/com/fsck/k9/view/MessageHeader.java
@@ -31,7 +31,7 @@ import com.fsck.k9.Account;
 import com.fsck.k9.FontSizes;
 import com.fsck.k9.K9;
 import com.fsck.k9.activity.misc.ContactPicture;
-import com.fsck.k9.activity.misc.ContactPictureLoader;
+import com.fsck.k9.contacts.ContactPictureLoader;
 import com.fsck.k9.ui.R;
 import com.fsck.k9.helper.ClipboardManager;
 import com.fsck.k9.helper.Contacts;
